+++ CHANGES.md
@@ -1,3 +1,7 @@
+# UNRELEASED
+
+- [FIXED] Set default value for `partitioned` parameter to false when creating a design document.
+
 # 2.13.0 (2020-04-16)
 
 - [FIXED] Correctly raise exceptions from `create_database` calls.